+++ web_joy_driver/CHANGELOG.rst
@@ -5,3 +5,4 @@ Changelog for package web_joy_driver
 Forthcoming
 -----------
 * feat: browser/smartphone virtual controller publishing sensor_msgs/Joy over WebSocket
+* feat: relay a sensor_msgs/CompressedImage camera topic to the page and show it between the sticks